refactor: moves validation logic back to core
The validation logic that exists works on top of the
trestle workspace in some cases.
