+++ trestle/core/links_validator.py
@@ -18,9 +18,9 @@ import logging
 import pathlib
 from typing import List, Optional, Set
 
-from trestle.common.common_types import TopLevelOscalModel
+from trestle.sdk.models.common_types import TopLevelOscalModel
 from trestle.common.model_utils import ModelUtils
-from trestle.sdk.validation.validator import Validator
+from trestle.core.validator import Validator
 
 logger = logging.getLogger(__name__)
