+++ trestle/core/refs_validator.py
@@ -17,9 +17,9 @@
 import pathlib
 from typing import Optional
 
-from trestle.common.common_types import TopLevelOscalModel
+from trestle.sdk.models.common_types import TopLevelOscalModel
 from trestle.common.list_utils import as_list
-from trestle.sdk.validation.validator import Validator
+from trestle.core.validator import Validator
 
 
 class RefsValidator(Validator):
